requirements.txt
# Base ----------------------------------------
pandas==1.4.2
python-dotenv==0.20.0 
requests==2.27.1

numpy>=1.18.5 
matplotlib>=3.2.2
seaborn>=0.11.2

# Machine learning ----------------------------
sklearn

tweepy==4.10.0